Introduced AbstractRange to eliminate the code duplications #2.
